+++ audionimbus/CHANGELOG.md
@@ -81,6 +81,7 @@
 - `OpenClDeviceLust::device_descriptor` now returns an `OpenClDeviceListError` error if the device index is out of bounds instead of panicking.
 - `channel` and `band` methods of `EnergyField` now return an `EnergyFieldError` error if arguments are out of bounds.
 - `Reconstructor::reconstruct` now returns a `ReconstructorError` instead of panicking, if the max duration or the max order is exceeded, or if the inputs and outputs have different lengths.
+- `ImpulseResponse::channel` now returns an `ImpulseResponseError` instead of panicking when the channel index is out of bounds.
 
 ### Added
 
@@ -98,7 +99,7 @@
 - Add `ChannelRequirement` to specify the channel count requirement for an audio buffer.
 - Add `Rendering` enum to choose between decoding ambisonics using binaural rendering or panning.
 - `SceneParams` implements the `Default` trait.
-- Add `ProbeArrayError`, `ProbeBatchErrors`, `OpenClDeviceListError`, `EnergyFieldError`, `ReconstructorError` errors.
+- Add `ProbeArrayError`, `ProbeBatchErrors`, `OpenClDeviceListError`, `EnergyFieldError`, `ReconstructorError`, `ImpulseResponseError` errors.
 
 ### Removed